Move LL stuff: clock
