+++ board/stm32fx/clock.h
